FilmsCatalog/requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.1
greenlet==1.1.2
idna==3.3
multidict==6.0.2
pip==21.1.2
python-dotenv==0.20.0
pytz==2022.2
requests==2.28.1
setuptools==57.0.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
urllib3==1.26.11
wheel==0.36.2
yarl==1.8.1
